Network.AWS.Snowball

Description

AWS Snowball is a petabyte-scale data transport solution that uses secure appliances to transfer large amounts of data between your on-premises data centers and Amazon Simple Storage Service (Amazon S3). The Snowball commands described here provide access to the same functionality that is available in the AWS Snowball Management Console, which enables you to create and manage jobs for Snowball. To transfer data locally with a Snowball appliance, you'll need to use the Snowball client or the Amazon S3 API adapter for Snowball. For more information, see the User Guide .

Service Configuration

snowball :: Service

API version 2016-06-30 of the Amazon Import/Export Snowball SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Snowball.

InvalidResourceException

_InvalidResourceException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource can't be found. Check the information you provided in your last request, and try again.

UnsupportedAddressException

_UnsupportedAddressException :: AsError a => Getting (First ServiceError) a ServiceError

The address is either outside the serviceable area for your region, or an error occurred. Check the address with your region's carrier and try again. If the issue persists, contact AWS Support.

KMSRequestFailedException

_KMSRequestFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The provided AWS Key Management Service key lacks the permissions to perform the specified CreateJob or UpdateJob action.

InvalidJobStateException

_InvalidJobStateException :: AsError a => Getting (First ServiceError) a ServiceError

The action can't be performed because the job's current state doesn't allow that action to be performed.

InvalidInputCombinationException

_InvalidInputCombinationException :: AsError a => Getting (First ServiceError) a ServiceError

Job or cluster creation failed. One ore more inputs were invalid. Confirm that the 'CreateClusterRequest$SnowballType' value supports your 'CreateJobRequest$JobType' , and try again.

InvalidNextTokenException

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The NextToken string was altered unexpectedly, and the operation has stopped. Run the operation without changing the NextToken string, and try again.

InvalidAddressException

_InvalidAddressException :: AsError a => Getting (First ServiceError) a ServiceError

The address provided was invalid. Check the address with your region's carrier, and try again.

ClusterLimitExceededException

_ClusterLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Job creation failed. Currently, clusters support five nodes. If you have less than five nodes for your cluster and you have more nodes to create for this cluster, try again and create jobs until your cluster has exactly five notes.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CancelCluster

module Network.AWS.Snowball.CancelCluster

DescribeCluster

module Network.AWS.Snowball.DescribeCluster

CreateAddress

module Network.AWS.Snowball.CreateAddress

GetSnowballUsage

module Network.AWS.Snowball.GetSnowballUsage

DescribeAddresses (Paginated)

module Network.AWS.Snowball.DescribeAddresses

UpdateCluster

module Network.AWS.Snowball.UpdateCluster

CreateJob

module Network.AWS.Snowball.CreateJob

GetJobManifest

module Network.AWS.Snowball.GetJobManifest

CreateCluster

module Network.AWS.Snowball.CreateCluster

ListJobs (Paginated)

module Network.AWS.Snowball.ListJobs

UpdateJob

module Network.AWS.Snowball.UpdateJob

GetJobUnlockCode

module Network.AWS.Snowball.GetJobUnlockCode

ListClusterJobs

module Network.AWS.Snowball.ListClusterJobs

DescribeJob

module Network.AWS.Snowball.DescribeJob

ListClusters

module Network.AWS.Snowball.ListClusters

DescribeAddress

module Network.AWS.Snowball.DescribeAddress

CancelJob

module Network.AWS.Snowball.CancelJob

Types

ClusterState

data ClusterState

Constructors

AwaitingQuorum

Cancelled

Complete

InUse

Pending

JobState

data JobState

Constructors

JSCancelled

JSComplete

JSInProgress

JSInTransitToAWS

JSInTransitToCustomer

JSListing

JSNew

JSPending

JSPreparingAppliance

JSPreparingShipment

JSWithAWS

JSWithCustomer

JobType

data JobType

Constructors

Export

Import

LocalUse

ShippingOption

data ShippingOption

Constructors

SOExpress

SONextDay

SOSecondDay

SOStandard

SnowballCapacity

data SnowballCapacity

Constructors

NoPreference

T100

T50

T80

SnowballType

data SnowballType

Constructors

Edge

Standard

Address

data Address

The address that you want the Snowball or Snowballs associated with a specific job to be shipped to. Addresses are validated at the time of creation. The address you provide must be located within the serviceable area of your region. Although no individual elements of the Address are required, if the address is invalid or unsupported, then an exception is thrown.

See: address smart constructor.

address :: Address

Creates a value of Address with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aIsRestricted - If the address you are creating is a primary address, then set this option to true. This field is not supported in most regions.

aStreet3 - The third line in a street address that a Snowball is to be delivered to.

aLandmark - This field is no longer used and the value is ignored.

aPostalCode - The postal code in an address that a Snowball is to be delivered to.

aCountry - The country in an address that a Snowball is to be delivered to.

aStateOrProvince - The state or province in an address that a Snowball is to be delivered to.

aStreet2 - The second line in a street address that a Snowball is to be delivered to.

aAddressId - The unique ID for an address.

aCity - The city in an address that a Snowball is to be delivered to.

aPhoneNumber - The phone number associated with an address that a Snowball is to be delivered to.

aCompany - The name of the company to receive a Snowball at an address.

aName - The name of a person to receive a Snowball at an address.

aPrefectureOrDistrict - This field is no longer used and the value is ignored.

aStreet1 - The first line in a street address that a Snowball is to be delivered to.

aIsRestricted :: Lens' Address (Maybe Bool)

If the address you are creating is a primary address, then set this option to true. This field is not supported in most regions.

aStreet3 :: Lens' Address (Maybe Text)

The third line in a street address that a Snowball is to be delivered to.

aLandmark :: Lens' Address (Maybe Text)

This field is no longer used and the value is ignored.

aPostalCode :: Lens' Address (Maybe Text)

The postal code in an address that a Snowball is to be delivered to.

aCountry :: Lens' Address (Maybe Text)

The country in an address that a Snowball is to be delivered to.

aStateOrProvince :: Lens' Address (Maybe Text)

The state or province in an address that a Snowball is to be delivered to.

aStreet2 :: Lens' Address (Maybe Text)

The second line in a street address that a Snowball is to be delivered to.

aAddressId :: Lens' Address (Maybe Text)

The unique ID for an address.

aCity :: Lens' Address (Maybe Text)

The city in an address that a Snowball is to be delivered to.

aPhoneNumber :: Lens' Address (Maybe Text)

The phone number associated with an address that a Snowball is to be delivered to.

aCompany :: Lens' Address (Maybe Text)

The name of the company to receive a Snowball at an address.

aName :: Lens' Address (Maybe Text)

The name of a person to receive a Snowball at an address.

aPrefectureOrDistrict :: Lens' Address (Maybe Text)

This field is no longer used and the value is ignored.

aStreet1 :: Lens' Address (Maybe Text)

The first line in a street address that a Snowball is to be delivered to.

ClusterListEntry

data ClusterListEntry

Contains a cluster's state, a cluster's ID, and other important information.

See: clusterListEntry smart constructor.

clusterListEntry :: ClusterListEntry

Creates a value of ClusterListEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cleClusterState - The current state of this cluster. For information about the state of a specific node, see 'JobListEntry$JobState' .

cleClusterId - The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000 .

cleCreationDate - The creation date for this cluster.

cleDescription - Defines an optional description of the cluster, for example Environmental Data Cluster-01 .

cleClusterState :: Lens' ClusterListEntry (Maybe ClusterState)

The current state of this cluster. For information about the state of a specific node, see 'JobListEntry$JobState' .

cleClusterId :: Lens' ClusterListEntry (Maybe Text)

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000 .

cleCreationDate :: Lens' ClusterListEntry (Maybe UTCTime)

The creation date for this cluster.

cleDescription :: Lens' ClusterListEntry (Maybe Text)

Defines an optional description of the cluster, for example Environmental Data Cluster-01 .

ClusterMetadata

data ClusterMetadata

Contains metadata about a specific cluster.

See: clusterMetadata smart constructor.

clusterMetadata :: ClusterMetadata

Creates a value of ClusterMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmJobType - The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE .

cmKMSKeyARN - The KmsKeyARN Amazon Resource Name (ARN) associated with this cluster. This ARN was created using the CreateKey API action in AWS Key Management Service (AWS KMS).

cmClusterState - The current status of the cluster.

cmNotification - The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

cmForwardingAddressId - The ID of the address that you want a cluster shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

cmAddressId - The automatically generated ID for a specific address.

cmSnowballType - The type of AWS Snowball appliance to use for this cluster. Currently, the only supported appliance type for cluster jobs is EDGE .

cmShippingOption - The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge appliance, rather it represents how quickly each appliance moves to its destination while in transit. Regional shipping speeds are as follows: * In Australia, you have access to express shipping. Typically, appliances shipped express are delivered in about a day. * In the European Union (EU), you have access to express shipping. Typically, Snowball Edges shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way. * In India, Snowball Edges are delivered in one to seven days. * In the US, you have access to one-day shipping and two-day shipping.

cmResources - The arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

cmClusterId - The automatically generated ID for a cluster.

cmCreationDate - The creation date for this cluster.

cmDescription - The optional description of the cluster.

cmRoleARN - The role ARN associated with this cluster. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

cmJobType :: Lens' ClusterMetadata (Maybe JobType)

The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE .

cmKMSKeyARN :: Lens' ClusterMetadata (Maybe Text)

The KmsKeyARN Amazon Resource Name (ARN) associated with this cluster. This ARN was created using the CreateKey API action in AWS Key Management Service (AWS KMS).

cmClusterState :: Lens' ClusterMetadata (Maybe ClusterState)

The current status of the cluster.

cmNotification :: Lens' ClusterMetadata (Maybe Notification)

The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

cmForwardingAddressId :: Lens' ClusterMetadata (Maybe Text)

The ID of the address that you want a cluster shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

cmAddressId :: Lens' ClusterMetadata (Maybe Text)

The automatically generated ID for a specific address.

cmSnowballType :: Lens' ClusterMetadata (Maybe SnowballType)

The type of AWS Snowball appliance to use for this cluster. Currently, the only supported appliance type for cluster jobs is EDGE .

cmShippingOption :: Lens' ClusterMetadata (Maybe ShippingOption)

The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge appliance, rather it represents how quickly each appliance moves to its destination while in transit. Regional shipping speeds are as follows: * In Australia, you have access to express shipping. Typically, appliances shipped express are delivered in about a day. * In the European Union (EU), you have access to express shipping. Typically, Snowball Edges shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way. * In India, Snowball Edges are delivered in one to seven days. * In the US, you have access to one-day shipping and two-day shipping.

cmResources :: Lens' ClusterMetadata (Maybe JobResource)

The arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

cmClusterId :: Lens' ClusterMetadata (Maybe Text)

The automatically generated ID for a cluster.

cmCreationDate :: Lens' ClusterMetadata (Maybe UTCTime)

The creation date for this cluster.

cmDescription :: Lens' ClusterMetadata (Maybe Text)

The optional description of the cluster.

cmRoleARN :: Lens' ClusterMetadata (Maybe Text)

The role ARN associated with this cluster. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

DataTransfer

data DataTransfer

Defines the real-time status of a Snowball's data transfer while the appliance is at AWS. This data is only available while a job has a JobState value of InProgress , for both import and export jobs.

See: dataTransfer smart constructor.

dataTransfer :: DataTransfer

Creates a value of DataTransfer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtTotalObjects - The total number of objects for a transfer between a Snowball and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dtTotalBytes - The total bytes of data for a transfer between a Snowball and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dtObjectsTransferred - The number of objects transferred between a Snowball and Amazon S3.

dtBytesTransferred - The number of bytes transferred between a Snowball and Amazon S3.

dtTotalObjects :: Lens' DataTransfer (Maybe Integer)

The total number of objects for a transfer between a Snowball and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dtTotalBytes :: Lens' DataTransfer (Maybe Integer)

The total bytes of data for a transfer between a Snowball and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dtObjectsTransferred :: Lens' DataTransfer (Maybe Integer)

The number of objects transferred between a Snowball and Amazon S3.

dtBytesTransferred :: Lens' DataTransfer (Maybe Integer)

The number of bytes transferred between a Snowball and Amazon S3.

EventTriggerDefinition

data EventTriggerDefinition

The container for the 'EventTriggerDefinition$EventResourceARN' .

See: eventTriggerDefinition smart constructor.

eventTriggerDefinition :: EventTriggerDefinition

Creates a value of EventTriggerDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etdEventResourceARN - The Amazon Resource Name (ARN) for any local Amazon S3 resource that is an AWS Lambda function's event trigger associated with this job.

etdEventResourceARN :: Lens' EventTriggerDefinition (Maybe Text)

The Amazon Resource Name (ARN) for any local Amazon S3 resource that is an AWS Lambda function's event trigger associated with this job.

JobListEntry

data JobListEntry

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of an export job.

See: jobListEntry smart constructor.

jobListEntry :: JobListEntry

Creates a value of JobListEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jleJobType - The type of job.

jleJobId - The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000 .

jleJobState - The current state of this job.

jleSnowballType - The type of appliance used with this job.

jleCreationDate - The creation date for this job.

jleDescription - The optional description of this specific job, for example Important Photos 2016-08-11 .

jleIsMaster - A value that indicates that this job is a master job. A master job represents a successful request to create an export job. Master jobs aren't associated with any Snowballs. Instead, each master job will have at least one job part, and each job part is associated with a Snowball. It might take some time before the job parts associated with a particular master job are listed, because they are created after the master job is created.

jleJobType :: Lens' JobListEntry (Maybe JobType)

The type of job.

jleJobId :: Lens' JobListEntry (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000 .

jleJobState :: Lens' JobListEntry (Maybe JobState)

The current state of this job.

jleSnowballType :: Lens' JobListEntry (Maybe SnowballType)

The type of appliance used with this job.

jleCreationDate :: Lens' JobListEntry (Maybe UTCTime)

The creation date for this job.

jleDescription :: Lens' JobListEntry (Maybe Text)

The optional description of this specific job, for example Important Photos 2016-08-11 .

jleIsMaster :: Lens' JobListEntry (Maybe Bool)

A value that indicates that this job is a master job. A master job represents a successful request to create an export job. Master jobs aren't associated with any Snowballs. Instead, each master job will have at least one job part, and each job part is associated with a Snowball. It might take some time before the job parts associated with a particular master job are listed, because they are created after the master job is created.

JobLogs

data JobLogs

Contains job logs. Whenever Snowball is used to import data into or export data out of Amazon S3, you'll have the option of downloading a PDF job report. Job logs are returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type. The job logs can be accessed for up to 60 minutes after this request has been made. To access any of the job logs after 60 minutes have passed, you'll have to make another call to the DescribeJob action.

For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snowball for your job part is being delivered to you.

The job report provides you insight into the state of your Amazon S3 data transfer. The report includes details about your job or job part for your records.

For deeper visibility into the status of your transferred objects, you can look at the two associated logs: a success log and a failure log. The logs are saved in comma-separated value (CSV) format, and the name of each log includes the ID of the job or job part that the log describes.

See: jobLogs smart constructor.

jobLogs :: JobLogs

Creates a value of JobLogs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jlJobFailureLogURI - A link to an Amazon S3 presigned URL where the job failure log is located.

jlJobCompletionReportURI - A link to an Amazon S3 presigned URL where the job completion report is located.

jlJobSuccessLogURI - A link to an Amazon S3 presigned URL where the job success log is located.

jlJobFailureLogURI :: Lens' JobLogs (Maybe Text)

A link to an Amazon S3 presigned URL where the job failure log is located.

jlJobCompletionReportURI :: Lens' JobLogs (Maybe Text)

A link to an Amazon S3 presigned URL where the job completion report is located.

jlJobSuccessLogURI :: Lens' JobLogs (Maybe Text)

A link to an Amazon S3 presigned URL where the job success log is located.

JobMetadata

data JobMetadata

Contains information about a specific job including shipping information, job status, and other important metadata. This information is returned as a part of the response syntax of the DescribeJob action.

See: jobMetadata smart constructor.

jobMetadata :: JobMetadata

Creates a value of JobMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jmJobType - The type of job.

jmKMSKeyARN - The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

jmJobId - The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000 .

jmJobLogInfo - Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snowball for your job part is being delivered to you.

jmNotification - The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

jmJobState - The current status of the jobs.

jmForwardingAddressId - The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

jmShippingDetails - A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

jmAddressId - The ID for the address that you want the Snowball shipped to.

jmSnowballType - The type of appliance used with this job.

jmDataTransferProgress - A value that defines the real-time status of a Snowball's data transfer while the appliance is at AWS. This data is only available while a job has a JobState value of InProgress , for both import and export jobs.

jmResources - An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

jmClusterId - The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000 .

jmCreationDate - The creation date for this job.

jmDescription - The description of the job, provided at job creation.

jmRoleARN - The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

jmSnowballCapacityPreference - The Snowball capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

jmJobType :: Lens' JobMetadata (Maybe JobType)

The type of job.

jmKMSKeyARN :: Lens' JobMetadata (Maybe Text)

The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

jmJobId :: Lens' JobMetadata (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000 .

jmJobLogInfo :: Lens' JobMetadata (Maybe JobLogs)

Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snowball for your job part is being delivered to you.

jmNotification :: Lens' JobMetadata (Maybe Notification)

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

jmJobState :: Lens' JobMetadata (Maybe JobState)

The current status of the jobs.

jmForwardingAddressId :: Lens' JobMetadata (Maybe Text)

The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

jmShippingDetails :: Lens' JobMetadata (Maybe ShippingDetails)

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

jmAddressId :: Lens' JobMetadata (Maybe Text)

The ID for the address that you want the Snowball shipped to.

jmSnowballType :: Lens' JobMetadata (Maybe SnowballType)

The type of appliance used with this job.

jmDataTransferProgress :: Lens' JobMetadata (Maybe DataTransfer)

A value that defines the real-time status of a Snowball's data transfer while the appliance is at AWS. This data is only available while a job has a JobState value of InProgress , for both import and export jobs.

jmResources :: Lens' JobMetadata (Maybe JobResource)

An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

jmClusterId :: Lens' JobMetadata (Maybe Text)

The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000 .

jmCreationDate :: Lens' JobMetadata (Maybe UTCTime)

The creation date for this job.

jmDescription :: Lens' JobMetadata (Maybe Text)

The description of the job, provided at job creation.

jmRoleARN :: Lens' JobMetadata (Maybe Text)

The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

jmSnowballCapacityPreference :: Lens' JobMetadata (Maybe SnowballCapacity)

The Snowball capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

JobResource

data JobResource

Contains an array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

See: jobResource smart constructor.

jobResource :: JobResource

Creates a value of JobResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jrLambdaResources - The Python-language Lambda functions for this job.

jrS3Resources - An array of S3Resource objects.

jrLambdaResources :: Lens' JobResource [LambdaResource]

The Python-language Lambda functions for this job.

jrS3Resources :: Lens' JobResource [S3Resource]

An array of S3Resource objects.

KeyRange

data KeyRange

Contains a key range. For export jobs, a S3Resource object can have an optional KeyRange value. The length of the range is defined at job creation, and has either an inclusive BeginMarker , an inclusive EndMarker , or both. Ranges are UTF-8 binary sorted.

See: keyRange smart constructor.

keyRange :: KeyRange

Creates a value of KeyRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

krEndMarker - The key that ends an optional key range for an export job. Ranges are inclusive and UTF-8 binary sorted.

krBeginMarker - The key that starts an optional key range for an export job. Ranges are inclusive and UTF-8 binary sorted.

krEndMarker :: Lens' KeyRange (Maybe Text)

The key that ends an optional key range for an export job. Ranges are inclusive and UTF-8 binary sorted.

krBeginMarker :: Lens' KeyRange (Maybe Text)

The key that starts an optional key range for an export job. Ranges are inclusive and UTF-8 binary sorted.

LambdaResource

data LambdaResource

Identifies

See: lambdaResource smart constructor.

lambdaResource :: LambdaResource

Creates a value of LambdaResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrEventTriggers - The array of ARNs for S3Resource objects to trigger the LambdaResource objects associated with this job.

lrLambdaARN - An Amazon Resource Name (ARN) that represents an AWS Lambda function to be triggered by PUT object actions on the associated local Amazon S3 resource.

lrEventTriggers :: Lens' LambdaResource [EventTriggerDefinition]

The array of ARNs for S3Resource objects to trigger the LambdaResource objects associated with this job.

lrLambdaARN :: Lens' LambdaResource (Maybe Text)

An Amazon Resource Name (ARN) that represents an AWS Lambda function to be triggered by PUT object actions on the associated local Amazon S3 resource.

Notification

data Notification

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

When the notification settings are defined during job creation, you can choose to notify based on a specific set of job states using the JobStatesToNotify array of strings, or you can specify that you want to have Amazon SNS notifications sent out for all job states with NotifyAll set to true.

See: notification smart constructor.

notification :: Notification

Creates a value of Notification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nNotifyAll - Any change in job state will trigger a notification for this job.

nSNSTopicARN - The new SNS TopicArn that you want to associate with this job. You can create Amazon Resource Names (ARNs) for topics by using the CreateTopic Amazon SNS API action. You can subscribe email addresses to an Amazon SNS topic through the AWS Management Console, or by using the Subscribe AWS Simple Notification Service (SNS) API action.

nJobStatesToNotify - The list of job states that will trigger a notification for this job.

nNotifyAll :: Lens' Notification (Maybe Bool)

Any change in job state will trigger a notification for this job.

nSNSTopicARN :: Lens' Notification (Maybe Text)

The new SNS TopicArn that you want to associate with this job. You can create Amazon Resource Names (ARNs) for topics by using the CreateTopic Amazon SNS API action. You can subscribe email addresses to an Amazon SNS topic through the AWS Management Console, or by using the Subscribe AWS Simple Notification Service (SNS) API action.

nJobStatesToNotify :: Lens' Notification [JobState]

The list of job states that will trigger a notification for this job.

S3Resource

data S3Resource

Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into. For export jobs, this object can have an optional KeyRange value. The length of the range is defined at job creation, and has either an inclusive BeginMarker , an inclusive EndMarker , or both. Ranges are UTF-8 binary sorted.

See: s3Resource smart constructor.

s3Resource :: S3Resource

Creates a value of S3Resource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srKeyRange - For export jobs, you can provide an optional KeyRange within a specific Amazon S3 bucket. The length of the range is defined at job creation, and has either an inclusive BeginMarker , an inclusive EndMarker , or both. Ranges are UTF-8 binary sorted.

srBucketARN - The Amazon Resource Name (ARN) of an Amazon S3 bucket.

srKeyRange :: Lens' S3Resource (Maybe KeyRange)

For export jobs, you can provide an optional KeyRange within a specific Amazon S3 bucket. The length of the range is defined at job creation, and has either an inclusive BeginMarker , an inclusive EndMarker , or both. Ranges are UTF-8 binary sorted.

srBucketARN :: Lens' S3Resource (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon S3 bucket.

Shipment

data Shipment

The Status and TrackingNumber information for an inbound or outbound shipment.

See: shipment smart constructor.

shipment :: Shipment

Creates a value of Shipment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sStatus - Status information for a shipment.

sTrackingNumber - The tracking number for this job. Using this tracking number with your region's carrier's website, you can track a Snowball as the carrier transports it. For India, the carrier is Amazon Logistics. For all other regions, UPS is the carrier.

sStatus :: Lens' Shipment (Maybe Text)

Status information for a shipment.

sTrackingNumber :: Lens' Shipment (Maybe Text)

The tracking number for this job. Using this tracking number with your region's carrier's website, you can track a Snowball as the carrier transports it. For India, the carrier is Amazon Logistics. For all other regions, UPS is the carrier.

ShippingDetails

data ShippingDetails

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

See: shippingDetails smart constructor.

shippingDetails :: ShippingDetails

Creates a value of ShippingDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdShippingOption - The shipping speed for a particular job. This speed doesn't dictate how soon you'll get the Snowball from the job's creation date. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows: * In Australia, you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. * In the European Union (EU), you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way. * In India, Snowballs are delivered in one to seven days. * In the United States of America (US), you have access to one-day shipping and two-day shipping.

sdOutboundShipment - The Status and TrackingNumber values for a Snowball being returned to AWS for a particular job.

sdInboundShipment - The Status and TrackingNumber values for a Snowball being delivered to the address that you specified for a particular job.

sdShippingOption :: Lens' ShippingDetails (Maybe ShippingOption)

The shipping speed for a particular job. This speed doesn't dictate how soon you'll get the Snowball from the job's creation date. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows: * In Australia, you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. * In the European Union (EU), you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way. * In India, Snowballs are delivered in one to seven days. * In the United States of America (US), you have access to one-day shipping and two-day shipping.

sdOutboundShipment :: Lens' ShippingDetails (Maybe Shipment)

The Status and TrackingNumber values for a Snowball being returned to AWS for a particular job.

sdInboundShipment :: Lens' ShippingDetails (Maybe Shipment)

The Status and TrackingNumber values for a Snowball being delivered to the address that you specified for a particular job.