Network.AWS.Snowball.CancelCluster

Description

Cancels a cluster job. You can only cancel a cluster job while it's in the AwaitingQuorum status. You'll have at least an hour after creating a cluster job to cancel it.

Creating a Request

cancelCluster

Arguments

:: Text
ccClusterId

-> CancelCluster

Creates a value of CancelCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccClusterId - The 39-character ID for the cluster that you want to cancel, for example CID123e4567-e89b-12d3-a456-426655440000 .

data CancelCluster

See: cancelCluster smart constructor.

Request Lenses

ccClusterId :: Lens' CancelCluster Text

The 39-character ID for the cluster that you want to cancel, for example CID123e4567-e89b-12d3-a456-426655440000 .

Destructuring the Response

cancelClusterResponse

Arguments

:: Int
ccrsResponseStatus

-> CancelClusterResponse

Creates a value of CancelClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccrsResponseStatus - -- | The response status code.

data CancelClusterResponse

See: cancelClusterResponse smart constructor.

Response Lenses

ccrsResponseStatus :: Lens' CancelClusterResponse Int

- | The response status code.