Network.AWS.Snowball.CreateAddress

Description

Creates an address for a Snowball to be shipped to. In most regions, addresses are validated at the time of creation. The address you provide must be located within the serviceable area of your region. If the address is invalid or unsupported, then an exception is thrown.

Creating a Request

createAddress

Arguments

:: Address
caAddress

-> CreateAddress

Creates a value of CreateAddress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caAddress - The address that you want the Snowball shipped to.

data CreateAddress

See: createAddress smart constructor.

Request Lenses

caAddress :: Lens' CreateAddress Address

The address that you want the Snowball shipped to.

Destructuring the Response

createAddressResponse

Arguments

:: Int
carsResponseStatus

-> CreateAddressResponse

Creates a value of CreateAddressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsAddressId - The automatically generated ID for a specific address. You'll use this ID when you create a job to specify which address you want the Snowball for that job shipped to.

carsResponseStatus - -- | The response status code.

data CreateAddressResponse

See: createAddressResponse smart constructor.

Response Lenses

carsAddressId :: Lens' CreateAddressResponse (Maybe Text)

The automatically generated ID for a specific address. You'll use this ID when you create a job to specify which address you want the Snowball for that job shipped to.

carsResponseStatus :: Lens' CreateAddressResponse Int

- | The response status code.