Network.AWS.Snowball.CreateJob

Description

Creates a job to import or export data between Amazon S3 and your on-premises data center. Your AWS account must have the right trust policies and permissions in place to create a job for Snowball. If you're creating a job for a node in a cluster, you only need to provide the clusterId value; the other job attributes are inherited from the cluster.

Creating a Request

createJob :: CreateJob

Creates a value of CreateJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjJobType - Defines the type of job that you're creating.

cjKMSKeyARN - The KmsKeyARN that you want to associate with this job. KmsKeyARN s are created using the CreateKey AWS Key Management Service (KMS) API action.

cjNotification - Defines the Amazon Simple Notification Service (Amazon SNS) notification settings for this job.

cjForwardingAddressId - The forwarding address ID for a job. This field is not supported in most regions.

cjAddressId - The ID for the address that you want the Snowball shipped to.

cjSnowballType - The type of AWS Snowball appliance to use for this job. Currently, the only supported appliance type for cluster jobs is EDGE .

cjShippingOption - The shipping speed for this job. This speed doesn't dictate how soon you'll get the Snowball, rather it represents how quickly the Snowball moves to its destination while in transit. Regional shipping speeds are as follows: * In Australia, you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. * In the European Union (EU), you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way. * In India, Snowballs are delivered in one to seven days. * In the US, you have access to one-day shipping and two-day shipping.

cjResources - Defines the Amazon S3 buckets associated with this job. With IMPORT jobs, you specify the bucket or buckets that your transferred data will be imported into. With EXPORT jobs, you specify the bucket or buckets that your transferred data will be exported from. Optionally, you can also specify a KeyRange value. If you choose to export a range, you define the length of the range by providing either an inclusive BeginMarker value, an inclusive EndMarker value, or both. Ranges are UTF-8 binary sorted.

cjClusterId - The ID of a cluster. If you're creating a job for a node in a cluster, you need to provide only this clusterId value. The other job attributes are inherited from the cluster.

cjDescription - Defines an optional description of this specific job, for example Important Photos 2016-08-11 .

cjRoleARN - The RoleARN that you want to associate with this job. RoleArn s are created using the CreateRole AWS Identity and Access Management (IAM) API action.

cjSnowballCapacityPreference - If your job is being created in one of the US regions, you have the option of specifying what size Snowball you'd like for this job. In all other regions, Snowballs come with 80 TB in storage capacity.

data CreateJob

See: createJob smart constructor.

Request Lenses

cjJobType :: Lens' CreateJob (Maybe JobType)

Defines the type of job that you're creating.

cjKMSKeyARN :: Lens' CreateJob (Maybe Text)

The KmsKeyARN that you want to associate with this job. KmsKeyARN s are created using the CreateKey AWS Key Management Service (KMS) API action.

cjNotification :: Lens' CreateJob (Maybe Notification)

Defines the Amazon Simple Notification Service (Amazon SNS) notification settings for this job.

cjForwardingAddressId :: Lens' CreateJob (Maybe Text)

The forwarding address ID for a job. This field is not supported in most regions.

cjAddressId :: Lens' CreateJob (Maybe Text)

The ID for the address that you want the Snowball shipped to.

cjSnowballType :: Lens' CreateJob (Maybe SnowballType)

The type of AWS Snowball appliance to use for this job. Currently, the only supported appliance type for cluster jobs is EDGE .

cjShippingOption :: Lens' CreateJob (Maybe ShippingOption)

The shipping speed for this job. This speed doesn't dictate how soon you'll get the Snowball, rather it represents how quickly the Snowball moves to its destination while in transit. Regional shipping speeds are as follows: * In Australia, you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. * In the European Union (EU), you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way. * In India, Snowballs are delivered in one to seven days. * In the US, you have access to one-day shipping and two-day shipping.

cjResources :: Lens' CreateJob (Maybe JobResource)

Defines the Amazon S3 buckets associated with this job. With IMPORT jobs, you specify the bucket or buckets that your transferred data will be imported into. With EXPORT jobs, you specify the bucket or buckets that your transferred data will be exported from. Optionally, you can also specify a KeyRange value. If you choose to export a range, you define the length of the range by providing either an inclusive BeginMarker value, an inclusive EndMarker value, or both. Ranges are UTF-8 binary sorted.

cjClusterId :: Lens' CreateJob (Maybe Text)

The ID of a cluster. If you're creating a job for a node in a cluster, you need to provide only this clusterId value. The other job attributes are inherited from the cluster.

cjDescription :: Lens' CreateJob (Maybe Text)

Defines an optional description of this specific job, for example Important Photos 2016-08-11 .

cjRoleARN :: Lens' CreateJob (Maybe Text)

The RoleARN that you want to associate with this job. RoleArn s are created using the CreateRole AWS Identity and Access Management (IAM) API action.

cjSnowballCapacityPreference :: Lens' CreateJob (Maybe SnowballCapacity)

If your job is being created in one of the US regions, you have the option of specifying what size Snowball you'd like for this job. In all other regions, Snowballs come with 80 TB in storage capacity.

Destructuring the Response

createJobResponse

Arguments

:: Int
cjrsResponseStatus

-> CreateJobResponse

Creates a value of CreateJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjrsJobId - The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000 .

cjrsResponseStatus - -- | The response status code.

data CreateJobResponse

See: createJobResponse smart constructor.

Response Lenses

cjrsJobId :: Lens' CreateJobResponse (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000 .

cjrsResponseStatus :: Lens' CreateJobResponse Int

- | The response status code.