Network.AWS.Snowball.DescribeAddresses

Description

Returns a specified number of ADDRESS objects. Calling this API in one of the US regions will return addresses from the list of all addresses associated with this account in all US regions.

This operation returns paginated results.

Creating a Request

describeAddresses :: DescribeAddresses

Creates a value of DescribeAddresses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daNextToken - HTTP requests are stateless. To identify what object comes "next" in the list of ADDRESS objects, you have the option of specifying a value for NextToken as the starting point for your list of returned addresses.

daMaxResults - The number of ADDRESS objects to return.

data DescribeAddresses

See: describeAddresses smart constructor.

Request Lenses

daNextToken :: Lens' DescribeAddresses (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of ADDRESS objects, you have the option of specifying a value for NextToken as the starting point for your list of returned addresses.

daMaxResults :: Lens' DescribeAddresses (Maybe Natural)

The number of ADDRESS objects to return.

Destructuring the Response

describeAddressesResponse

Arguments

:: Int
drsResponseStatus

-> DescribeAddressesResponse

Creates a value of DescribeAddressesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsAddresses - The Snowball shipping addresses that were created for this account.

drsNextToken - HTTP requests are stateless. If you use the automatically generated NextToken value in your next DescribeAddresses call, your list of returned addresses will start from this point in the array.

drsResponseStatus - -- | The response status code.

data DescribeAddressesResponse

See: describeAddressesResponse smart constructor.

Response Lenses

drsAddresses :: Lens' DescribeAddressesResponse [Address]

The Snowball shipping addresses that were created for this account.

drsNextToken :: Lens' DescribeAddressesResponse (Maybe Text)

HTTP requests are stateless. If you use the automatically generated NextToken value in your next DescribeAddresses call, your list of returned addresses will start from this point in the array.

drsResponseStatus :: Lens' DescribeAddressesResponse Int

- | The response status code.