Network.AWS.Snowball.ListClusterJobs

Description

Returns an array of JobListEntry objects of the specified length. Each JobListEntry object is for a job in the specified cluster and contains a job's state, a job's ID, and other information.

Creating a Request

listClusterJobs

Arguments

:: Text
lcjClusterId

-> ListClusterJobs

Creates a value of ListClusterJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcjNextToken - HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

lcjMaxResults - The number of JobListEntry objects to return.

lcjClusterId - The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000 .

data ListClusterJobs

See: listClusterJobs smart constructor.

Request Lenses

lcjNextToken :: Lens' ListClusterJobs (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

lcjMaxResults :: Lens' ListClusterJobs (Maybe Natural)

The number of JobListEntry objects to return.

lcjClusterId :: Lens' ListClusterJobs Text

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000 .

Destructuring the Response

listClusterJobsResponse

Arguments

:: Int
lcjrsResponseStatus

-> ListClusterJobsResponse

Creates a value of ListClusterJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcjrsJobListEntries - Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

lcjrsNextToken - HTTP requests are stateless. If you use the automatically generated NextToken value in your next ListClusterJobsResult call, your list of returned jobs will start from this point in the array.

lcjrsResponseStatus - -- | The response status code.

data ListClusterJobsResponse

See: listClusterJobsResponse smart constructor.

Response Lenses

lcjrsJobListEntries :: Lens' ListClusterJobsResponse [JobListEntry]

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

lcjrsNextToken :: Lens' ListClusterJobsResponse (Maybe Text)

HTTP requests are stateless. If you use the automatically generated NextToken value in your next ListClusterJobsResult call, your list of returned jobs will start from this point in the array.

lcjrsResponseStatus :: Lens' ListClusterJobsResponse Int

- | The response status code.