Network.AWS.Snowball.ListJobs

Description

Returns an array of JobListEntry objects of the specified length. Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs. Calling this API action in one of the US regions will return jobs from the list of all jobs associated with this account in all US regions.

This operation returns paginated results.

Creating a Request

listJobs :: ListJobs

Creates a value of ListJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ljNextToken - HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

ljMaxResults - The number of JobListEntry objects to return.

data ListJobs

See: listJobs smart constructor.

Request Lenses

ljNextToken :: Lens' ListJobs (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

ljMaxResults :: Lens' ListJobs (Maybe Natural)

The number of JobListEntry objects to return.

Destructuring the Response

listJobsResponse

Arguments

:: Int
ljrsResponseStatus

-> ListJobsResponse

Creates a value of ListJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ljrsJobListEntries - Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

ljrsNextToken - HTTP requests are stateless. If you use this automatically generated NextToken value in your next ListJobs call, your returned JobListEntry objects will start from this point in the array.

ljrsResponseStatus - -- | The response status code.

data ListJobsResponse

See: listJobsResponse smart constructor.

Response Lenses

ljrsJobListEntries :: Lens' ListJobsResponse [JobListEntry]

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

ljrsNextToken :: Lens' ListJobsResponse (Maybe Text)

HTTP requests are stateless. If you use this automatically generated NextToken value in your next ListJobs call, your returned JobListEntry objects will start from this point in the array.

ljrsResponseStatus :: Lens' ListJobsResponse Int

- | The response status code.