Network.AWS.Snowball.UpdateCluster

Description

While a cluster's ClusterState value is in the AwaitingQuorum state, you can update some of the information associated with a cluster. Once the cluster changes to a different job state, usually 60 minutes after the cluster being created, this action is no longer available.

Creating a Request

updateCluster

Arguments

:: Text
ucClusterId

-> UpdateCluster

Creates a value of UpdateCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucNotification - The new or updated Notification object.

ucForwardingAddressId - The updated ID for the forwarding address for a cluster. This field is not supported in most regions.

ucAddressId - The ID of the updated Address object.

ucShippingOption - The updated shipping option value of this cluster's ShippingDetails object.

ucResources - The updated arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

ucDescription - The updated description of this cluster.

ucRoleARN - The new role Amazon Resource Name (ARN) that you want to associate with this cluster. To create a role ARN, use the CreateRole API action in AWS Identity and Access Management (IAM).

ucClusterId - The cluster ID of the cluster that you want to update, for example CID123e4567-e89b-12d3-a456-426655440000 .

data UpdateCluster

See: updateCluster smart constructor.

Request Lenses

ucNotification :: Lens' UpdateCluster (Maybe Notification)

The new or updated Notification object.

ucForwardingAddressId :: Lens' UpdateCluster (Maybe Text)

The updated ID for the forwarding address for a cluster. This field is not supported in most regions.

ucAddressId :: Lens' UpdateCluster (Maybe Text)

The ID of the updated Address object.

ucShippingOption :: Lens' UpdateCluster (Maybe ShippingOption)

The updated shipping option value of this cluster's ShippingDetails object.

ucResources :: Lens' UpdateCluster (Maybe JobResource)

The updated arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

ucDescription :: Lens' UpdateCluster (Maybe Text)

The updated description of this cluster.

ucRoleARN :: Lens' UpdateCluster (Maybe Text)

The new role Amazon Resource Name (ARN) that you want to associate with this cluster. To create a role ARN, use the CreateRole API action in AWS Identity and Access Management (IAM).

ucClusterId :: Lens' UpdateCluster Text

The cluster ID of the cluster that you want to update, for example CID123e4567-e89b-12d3-a456-426655440000 .

Destructuring the Response

updateClusterResponse

Arguments

:: Int
ucrsResponseStatus

-> UpdateClusterResponse

Creates a value of UpdateClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucrsResponseStatus - -- | The response status code.

data UpdateClusterResponse

See: updateClusterResponse smart constructor.

Response Lenses

ucrsResponseStatus :: Lens' UpdateClusterResponse Int

- | The response status code.