Network.AWS.Snowball.UpdateJob

Description

While a job's JobState value is New , you can update some of the information associated with a job. Once the job changes to a different job state, usually within 60 minutes of the job being created, this action is no longer available.

Creating a Request

updateJob

Arguments

:: Text
ujJobId

-> UpdateJob

Creates a value of UpdateJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ujNotification - The new or updated Notification object.

ujForwardingAddressId - The updated ID for the forwarding address for a job. This field is not supported in most regions.

ujAddressId - The ID of the updated Address object.

ujShippingOption - The updated shipping option value of this job's ShippingDetails object.

ujResources - The updated S3Resource object (for a single Amazon S3 bucket or key range), or the updated JobResource object (for multiple buckets or key ranges).

ujDescription - The updated description of this job's JobMetadata object.

ujRoleARN - The new role Amazon Resource Name (ARN) that you want to associate with this job. To create a role ARN, use the CreateRole AWS Identity and Access Management (IAM) API action.

ujSnowballCapacityPreference - The updated SnowballCapacityPreference of this job's JobMetadata object. The 50 TB Snowballs are only available in the US regions.

ujJobId - The job ID of the job that you want to update, for example JID123e4567-e89b-12d3-a456-426655440000 .

data UpdateJob

See: updateJob smart constructor.

Request Lenses

ujNotification :: Lens' UpdateJob (Maybe Notification)

The new or updated Notification object.

ujForwardingAddressId :: Lens' UpdateJob (Maybe Text)

The updated ID for the forwarding address for a job. This field is not supported in most regions.

ujAddressId :: Lens' UpdateJob (Maybe Text)

The ID of the updated Address object.

ujShippingOption :: Lens' UpdateJob (Maybe ShippingOption)

The updated shipping option value of this job's ShippingDetails object.

ujResources :: Lens' UpdateJob (Maybe JobResource)

The updated S3Resource object (for a single Amazon S3 bucket or key range), or the updated JobResource object (for multiple buckets or key ranges).

ujDescription :: Lens' UpdateJob (Maybe Text)

The updated description of this job's JobMetadata object.

ujRoleARN :: Lens' UpdateJob (Maybe Text)

The new role Amazon Resource Name (ARN) that you want to associate with this job. To create a role ARN, use the CreateRole AWS Identity and Access Management (IAM) API action.

ujSnowballCapacityPreference :: Lens' UpdateJob (Maybe SnowballCapacity)

The updated SnowballCapacityPreference of this job's JobMetadata object. The 50 TB Snowballs are only available in the US regions.

ujJobId :: Lens' UpdateJob Text

The job ID of the job that you want to update, for example JID123e4567-e89b-12d3-a456-426655440000 .

Destructuring the Response

updateJobResponse

Arguments

:: Int
ujrsResponseStatus

-> UpdateJobResponse

Creates a value of UpdateJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ujrsResponseStatus - -- | The response status code.

data UpdateJobResponse

See: updateJobResponse smart constructor.

Response Lenses

ujrsResponseStatus :: Lens' UpdateJobResponse Int

- | The response status code.